refactor(v2): move session storage under storage
