+++ packages/opencode/test/session/prompt.test.ts
@@ -36,7 +36,7 @@ import { SessionRunState } from "../../src/session/run-state"
 import { MessageID, PartID, SessionID } from "../../src/session/schema"
 import { SessionStatus } from "../../src/session/status"
 import { SessionV2 } from "../../src/v2/session"
-import { SessionStorageSql } from "../../src/v2/session/storage-sql"
+import { SessionStorageSql } from "../../src/v2/storage/session-sql"
 import { Skill } from "../../src/skill"
 import { SystemPrompt } from "../../src/session/system"
 import { Shell } from "../../src/shell/shell"
