+++ packages/opencode/test/v2/session-storage.test.ts
@@ -4,9 +4,9 @@ import { ProjectTable } from "@/project/project.sql"
 import { SessionID } from "@/session/schema"
 import { SessionMessageTable, SessionTable } from "@/session/session.sql"
 import { Database } from "@/storage/db"
-import { SessionStorage } from "@/v2/session/storage"
-import { SessionStorageMemory } from "@/v2/session/storage-memory"
-import { SessionStorageSql } from "@/v2/session/storage-sql"
+import { SessionStorage } from "@/v2/storage/session"
+import { SessionStorageMemory } from "@/v2/storage/session-memory"
+import { SessionStorageSql } from "@/v2/storage/session-sql"
 import { EventV2 } from "@opencode-ai/core/event"
 import { SessionMessage } from "@opencode-ai/core/session-message"
 import { eq, or } from "@/storage/db"
